+++ travharv/__init__.py
@@ -5,16 +5,16 @@
 :synopsis: A package for traversing and harvesting RDF data
 by dereferencing URIs and asserting paths.
 
-.. moduleauthor:: Cedric Decruw <cedric.decruw@vliz.be>
+.. moduleauthor:: "Flanders Marine Institute, VLIZ vzw" <opsci@vliz.be>
 """
 
 from travharv.service import TravHarv
 from travharv.store import TargetStoreAccess as RDFStoreAccess
-from travharv.trav_harv_config_builder import (
+from travharv.config_build import (
     TravHarvConfig,
     TravHarvConfigBuilder,
 )
-from travharv.trav_harv_executer import TravHarvExecutor
+from travharv.executor import TravHarvExecutor
 
 __all__ = [
     "RDFStoreAccess",
