+++ travharv/__main__.py
@@ -11,11 +11,11 @@ from rdflib import Graph
 
 from travharv.common import QUERY_BUILDER, insert_resource_into_graph
 from travharv.store import TargetStoreAccess as RDFStoreAccess
-from travharv.trav_harv_config_builder import (
+from travharv.config_build import (
     TravHarvConfig,
     TravHarvConfigBuilder,
 )
-from travharv.trav_harv_executer import TravHarvExecutor
+from travharv.executor import TravHarvExecutor
 
 # log = logging.getLogger("travharv")
 log = logging.getLogger(__name__)
